# perf(options): refactor how options are sanitized