… spec/extendObject.spec.ts
… src/utils.ts