Related to #43

* Fixes bug where `schema` would throw when passed `undefined`
* Adds tests for passing `undefined` for every options property
* Removes the need for `extendObject`
* This change results in the following benchmark results below. TL;DR: faster in every scenario than the previous release by millions of operations per second and bundle size is overall reduced.

```
                    Asset       Size  Chunks             Chunk Names
    isDatatype.umd.min.js    4.75 kB       0  [emitted]  isDatatype
isDatatype.umd.min.js.map  125 bytes       0  [emitted]  isDatatype
 isDatatype.umd.min.js.gz     1.8 kB          [emitted]  
   [0] ../index.ts + 6 modules 13.2 kB {0} [built]

                Asset     Size  Chunks             Chunk Names
    isDatatype.umd.js  10.3 kB       0  [emitted]  isDatatype
isDatatype.umd.js.map  1.08 kB       0  [emitted]  isDatatype
   [0] ../index.ts + 6 modules 13.2 kB {0} [built]

> is-datatype@0.3.1 benchmark /Users/emilio.martinez/Documents/emilio/is-datatype
> node tools/benchmarks/benchmarks.js

[Any] v0.3.1 x 1,296,242 ops/sec ±0.64% (289 runs sampled)
[Any] latest x 50,054,882 ops/sec ±0.73% (288 runs sampled)
[Any] Fastest is 'latest' (3,862%)

[Undefined (valid)] v0.3.1 x 1,389,232 ops/sec ±0.62% (289 runs sampled)
[Undefined (valid)] latest x 6,779,993 ops/sec ±0.62% (289 runs sampled)
[Undefined (valid)] Fastest is 'latest' (488%)

[Undefined (invalid)] v0.3.1 x 1,372,843 ops/sec ±0.61% (289 runs sampled)
[Undefined (invalid)] latest x 6,531,763 ops/sec ±0.58% (288 runs sampled)
[Undefined (invalid)] Fastest is 'latest' (476%)

[Boolean (valid)] v0.3.1 x 1,362,019 ops/sec ±0.57% (287 runs sampled)
[Boolean (valid)] latest x 6,551,905 ops/sec ±0.59% (288 runs sampled)
[Boolean (valid)] Fastest is 'latest' (481%)

[Boolean (invalid)] v0.3.1 x 1,367,964 ops/sec ±0.58% (287 runs sampled)
[Boolean (invalid)] latest x 6,617,794 ops/sec ±0.57% (288 runs sampled)
[Boolean (invalid)] Fastest is 'latest' (484%)

[Number (valid)] v0.3.1 x 1,295,890 ops/sec ±0.67% (288 runs sampled)
[Number (valid)] latest x 6,064,311 ops/sec ±0.63% (288 runs sampled)
[Number (valid)] Fastest is 'latest' (468%)

[Number (invalid)] v0.3.1 x 1,382,457 ops/sec ±0.47% (289 runs sampled)
[Number (invalid)] latest x 6,780,928 ops/sec ±0.58% (288 runs sampled)
[Number (invalid)] Fastest is 'latest' (490%)

[Integer (valid)] v0.3.1 x 116,708 ops/sec ±0.52% (286 runs sampled)
[Integer (valid)] latest x 3,591,586 ops/sec ±0.51% (290 runs sampled)
[Integer (valid)] Fastest is 'latest' (3,077%)

[Integer (invalid)] v0.3.1 x 117,909 ops/sec ±0.49% (287 runs sampled)
[Integer (invalid)] latest x 3,949,169 ops/sec ±0.62% (289 runs sampled)
[Integer (invalid)] Fastest is 'latest' (3,349%)

[Natural (valid)] v0.3.1 x 110,317 ops/sec ±0.64% (286 runs sampled)
[Natural (valid)] latest x 3,627,236 ops/sec ±0.55% (289 runs sampled)
[Natural (valid)] Fastest is 'latest' (3,288%)

[Natural (invalid)] v0.3.1 x 120,056 ops/sec ±0.67% (287 runs sampled)
[Natural (invalid)] latest x 3,979,628 ops/sec ±0.55% (288 runs sampled)
[Natural (invalid)] Fastest is 'latest' (3,315%)

[String (valid)] v0.3.1 x 1,051,555 ops/sec ±0.63% (289 runs sampled)
[String (valid)] latest x 2,940,575 ops/sec ±0.51% (287 runs sampled)
[String (valid)] Fastest is 'latest' (280%)

[String (invalid)] v0.3.1 x 1,348,398 ops/sec ±0.60% (288 runs sampled)
[String (invalid)] latest x 6,203,468 ops/sec ±0.72% (285 runs sampled)
[String (invalid)] Fastest is 'latest' (460%)

[Function (valid)] v0.3.1 x 1,347,702 ops/sec ±0.52% (289 runs sampled)
[Function (valid)] latest x 6,076,122 ops/sec ±0.87% (284 runs sampled)
[Function (valid)] Fastest is 'latest' (451%)

[Function (invalid)] v0.3.1 x 1,345,779 ops/sec ±0.61% (289 runs sampled)
[Function (invalid)] latest x 6,167,116 ops/sec ±0.84% (286 runs sampled)
[Function (invalid)] Fastest is 'latest' (458%)

[Object (valid)] v0.3.1 x 1,293,168 ops/sec ±0.56% (289 runs sampled)
[Object (valid)] latest x 5,793,900 ops/sec ±0.67% (288 runs sampled)
[Object (valid)] Fastest is 'latest' (448%)

[Object (invalid)] v0.3.1 x 1,342,833 ops/sec ±0.59% (288 runs sampled)
[Object (invalid)] latest x 6,653,659 ops/sec ±0.55% (289 runs sampled)
[Object (invalid)] Fastest is 'latest' (495%)

[Array (valid)] v0.3.1 x 1,192,815 ops/sec ±0.65% (287 runs sampled)
[Array (valid)] latest x 5,185,396 ops/sec ±0.57% (288 runs sampled)
[Array (valid)] Fastest is 'latest' (435%)

[Array (invalid)] v0.3.1 x 1,335,401 ops/sec ±0.87% (289 runs sampled)
[Array (invalid)] latest x 5,795,722 ops/sec ±8.04% (282 runs sampled)
[Array (invalid)] Fastest is 'latest' (434%)
```